+++ src/core/events/sensor/activated/sensor-activated-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorActivated extends SensorEvent {
   static version = '1';
