+++ src/core/events/sensor/deactivated/sensor-deactivated-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorDeactivated extends SensorEvent {
   static version = '1';
