+++ src/core/events/sensor/deleted/index.ts
@@ -1,8 +1,8 @@
 import { plainToClass } from 'class-transformer';
-import { SensorDeleted as V1 } from './1/sensor-deleted.event';
+import { SensorDeleted as V1 } from './sensor-deleted-v1.event';
 import { EventMessage } from '../../../../event-store/event-message';
 
-export { SensorDeleted } from './1/sensor-deleted.event';
+export { SensorDeleted } from './sensor-deleted-v1.event';
 
 export function getSensorDeletedEvent(eventMessage: EventMessage): V1 {
     return !eventMessage.metadata.version || eventMessage.metadata.version === V1.version ? plainToClass(V1, eventMessage.data) : null;
