+++ src/core/events/sensor/deleted/sensor-deleted-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorDeleted extends SensorEvent {
   static version = '1';
