+++ src/core/events/sensor/ds-deleted/index.ts
@@ -1,8 +1,8 @@
 import { plainToClass } from 'class-transformer';
 import { EventMessage } from '../../../../event-store/event-message';
-import { DatastreamDeleted as V1 } from './1/sensor-datastreamdeleted.event';
+import { DatastreamDeleted as V1 } from './sensor-datastreamdeleted-v1.event';
 
-export { DatastreamDeleted } from './1/sensor-datastreamdeleted.event';
+export { DatastreamDeleted } from './sensor-datastreamdeleted-v1.event';
 
 export function getDatastreamDeletedEvent(eventMessage: EventMessage): V1 {
     return !eventMessage.metadata.version || eventMessage.metadata.version === V1.version ? plainToClass(V1, eventMessage.data) : null;
