+++ src/core/events/sensor/ds-deleted/sensor-datastreamdeleted-v1.event.ts
@@ -1,4 +1,4 @@
-import {SensorEvent} from '../../sensor.event';
+import {SensorEvent} from '../sensor.event';
 
 export class DatastreamDeleted extends SensorEvent {
   static version = '1';
