--- src/core/events/sensor/ds-updated/1/sensor-datastreamupdated.event.ts
@@ -1,3 +1,0 @@
-import { DatastreamChanged } from '../../sensor-datastreamchanged.event';
-
-export class DatastreamUpdated extends DatastreamChanged {}
