+++ src/core/events/sensor/ownership-shared/sensor-ownershipshared-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorOwnershipShared extends SensorEvent {
   static version = '1';
