+++ src/core/events/sensor/ownership-transferred/sensor-ownershiptransferred-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorOwnershipTransferred extends SensorEvent {
   static version = '1';
