+++ src/core/events/sensor/registered/sensor-registered-v1.event.ts
@@ -1,11 +1,13 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorRegistered extends SensorEvent {
-  static version = '2';
+  static version = '1';
 
   readonly organizationId: string;
   readonly name: string;
-  readonly location: number[];
+  readonly longitude: number;
+  readonly latitude: number;
+  readonly height: number;
   readonly baseObjectId: string;
   readonly aim: string;
   readonly description: string;
@@ -18,14 +20,18 @@ export class SensorRegistered extends SensorEvent {
   readonly typeName: string;
   readonly typeDetails: Record<string, any>;
 
-  constructor(sensorId: string, organizationId: string, name: string, location: number[], baseObjectId: string,
-              aim: string, description: string, manufacturer: string, active: boolean,
-              observationArea: Record<string, any>, documentationUrl: string, theme: string[], category: string,
-              typeName: string, typeDetails: Record<string, any>) {
+  constructor(sensorId: string, organizationId: string,
+              name: string, longitude: number, latitude: number, height: number,
+              baseObjectId: string, aim: string, description: string,
+              manufacturer: string, active: boolean, observationArea: Record<string, any>,
+              documentationUrl: string, theme: string[], category: string, typeName: string,
+              typeDetails: Record<string, any>) {
     super(sensorId, SensorRegistered.version);
     this.organizationId = organizationId;
     this.name = name;
-    this.location = location;
+    this.longitude = longitude;
+    this.latitude = latitude;
+    this.height = height;
     this.baseObjectId = baseObjectId;
     this.aim = aim;
     this.description = description;
