+++ src/core/events/sensor/relocated/sensor-relocated-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorRelocated extends SensorEvent {
   static version = '1';
