+++ src/core/events/sensor/updated/sensor-updated-v1.event.ts
@@ -1,4 +1,4 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 export class SensorUpdated extends SensorEvent {
   static version = '1';
